requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch==2.4.1
sentencepiece==0.2.0
huggingface_hub==0.24.6
rapidfuzz==3.9.7
pydantic==2.8.2
accelerate==0.34.2
python-dotenv==1.0.1
requests==2.32.3
faster-whisper==1.0.3
pyttsx3==2.99
gTTS==2.5.4
